Прототип ui компаса

+++ app/src/main/java/ru/upsoft/gpspointer/domain/model/CompassState.kt
@@ -1,6 +1,6 @@
 package ru.upsoft.gpspointer.domain.model
 
 sealed interface CompassState {
-    data class Loaded(val degree: Double) : CompassState
+    data class Loaded(val degree: Float) : CompassState
 }
